// data/github/current/ui/packages/issue-create/__tests__/__generated__/CreateIssueTestQuery.graphql.ts
export type CreateIssueTestQuery$variables = Record<PropertyKey, never>;
export type CreateIssueTestQuery$data = {
  readonly repository: {
    readonly " $fragmentSpreads": FragmentRefs<"CreateIssue">;
  } | null | undefined;
};
export type CreateIssueTestQuery = {
  response: CreateIssueTestQuery$data;
  variables: CreateIssueTestQuery$variables;
};